Added begin and commit support

+++ pa4/jesql_commands.py
@@ -75,7 +75,10 @@ def drop_table(name):
 
     # check if table exists and remove the file
     if os.path.exists(database_dir + "/" + name):
-        os.remove(database_dir + "/" + name)
+        if not os.path.exists(database_dir, + "/" + name + ".lock"):
+            os.remove(database_dir + "/" + name)
+        else:
+            print("!Failed to delete", name, "because it currently in use.", file=sys.stderr)
     else:
         print ("!Failed to delete", name, "because it does not exist.")
 
@@ -286,3 +289,9 @@ def update(tbname, conditional, set_attr, set_val, where_attr, where_val):
         print(records_updated, 'records modified.')
     except FileNotFoundError:
         print('ERROR: Invalid table name')
+
+def begin_transaction():
+    return true;
+
+def commit():
+    return false;
